+++ src/library/math/operator/exports.ts
@@ -2,4 +2,4 @@ export * from './predicate';
 
 export * from './hyper';
 
-export * from './constructive';
+export * from './constructive/exports';
